otpl: move code/rename files
